adding 1d continuum solution and minor
